requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.2.5
parsel==1.5.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.8.0
python-dotenv==0.10.1
PyYAML==5.1.2
queuelib==1.5.0
requests==2.21.0
rsa==4.0
Scrapy==1.5.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
spotipy==2.4.4
Twisted==19.7.0
uritemplate==3.0.0
urllib3==1.24.2
w3lib==1.19.0
zope.interface==4.6.0
